// src/infrastructure/persistence/enums.ts
export enum StudySessionMode {
  FLASHCARD = 'FLASHCARD',
  LEARN = 'LEARN',
  TEST = 'TEST',
  MATCH = 'MATCH',
}

export enum CollaboratorRole {
  OWNER = 'OWNER',
  EDITOR = 'EDITOR',
  VIEWER = 'VIEWER',
}

export enum StudySetVisibility {
  PUBLIC = 'PUBLIC',
  PRIVATE = 'PRIVATE',
  UNLISTED = 'UNLISTED',
}